Move linefinder to finder to account for extrema and fast
