+++ gaiaxpy/__init__.py
@@ -6,6 +6,6 @@ from .core.version import __version__
 from .error_correction.error_correction import apply_error_correction
 from .generator.generator import generate
 from .generator.photometric_system import PhotometricSystem, load_additional_systems, remove_additional_systems
-from .lines.linefinder import extremafinder, fastfinder, linefinder
+from .finder.linefinder import extremafinder, fastfinder, linefinder
 from .plotter.plot_spectra import plot_spectra
