+++ tests/test_lines/test_fastfinder.py
@@ -4,7 +4,7 @@ from os.path import join
 import pandas as pd
 import pandas.testing as pdt
 
-from gaiaxpy.lines.linefinder import fastfinder
+from gaiaxpy.finder.linefinder import fastfinder
 from tests.files.paths import files_path
 from tests.utils.utils import get_converters
 
@@ -13,7 +13,7 @@ _rtol, _atol = 1e-7, 1e-7
 dtypes = [('line_name', 'U12'), ('wavelength_nm', 'f8'), ('flux', 'f8'), ('depth', 'f8'), ('width', 'f8'),
           ('significance', 'f8'), ('sig_pwl', 'f8')]
 
-solution_folder = 'lines_files'
+solution_folder = 'finder_files'
 found_fast_real = pd.read_csv(join(files_path, solution_folder, 'fastfinder_output.csv'),
                               converters=get_converters(['extrema_bp', 'extrema_rp']))
 found_fast_trunc_real = pd.read_csv(join(files_path, solution_folder, 'fastfinder_trunc_output.csv'),
