+++ tests/test_lines/test_herm.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 import numpy.testing as npt
 
-from gaiaxpy.lines.herm import HermiteDerivative
+from gaiaxpy.finder.herm import HermiteDerivative
 
 n = 4
 tm = np.ones((n, n))
